refactor: Migrate GoRouter to new API, enhance resource detail description formatting with regex, and update dependencies.

+++ lib/app/home/onboarding/onboarding_carusel.dart
@@ -169,7 +169,7 @@ class _OnboardingCarouselState extends State<OnboardingCarousel> {
   }) {
     return DotsIndicator(
       dotsCount: pageLength,
-      position: currentIndex,
+      position: currentIndex.toDouble(),
       onTap: (index) {
         _moveToNextCarousel(index.toInt());
       },
